+++ SickSangHae/Application/Assets.xcassets/BlueGray.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0.988",
-          "green" : "0.980",
-          "red" : "0.965"
+          "blue" : "0.714",
+          "green" : "0.643",
+          "red" : "0.553"
         }
       },
       "idiom" : "universal"
