+++ SickSangHae/Application/Assets.xcassets/LightBlueGray.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0.949",
-          "green" : "0.949",
-          "red" : "0.949"
+          "blue" : "0.871",
+          "green" : "0.839",
+          "red" : "0.788"
         }
       },
       "idiom" : "universal"
