+++ SickSangHae/Application/Assets.xcassets/LightGreenGray.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0.973",
-          "green" : "0.969",
-          "red" : "0.969"
+          "blue" : "0.949",
+          "green" : "0.941",
+          "red" : "0.800"
         }
       },
       "idiom" : "universal"
